routes in progress

+++ app/Http/Controllers/AuthLoginRegisterController.php
@@ -104,12 +104,15 @@ class AuthLoginRegisterController extends Controller
 
     /* Logout Revoca el token de acceso del usuario actual, 
     lo desconecta y devuelve una respuesta en json.*/
-    
+
     public function logout(Request $request)
     {
-        $request->user()->currentAccessToken()->delete();
-
-        return response()->json(['message' => 'Logged out successfully'], 200);
+        if ($request->user()) {
+            $request->user()->currentAccessToken()->delete();
+            return response()->json(['message' => 'Logged out successfully'], 200);
+        } else {
+            return response()->json(['message' => 'No user authenticated'], 404);
+        }
     }
 
     // public function user(Request $request)
